requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==8.6.1
isort==6.0.0
itsdangerous==2.2.0
jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.17
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langgraph==0.2.69
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.4
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.61.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.1
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
tenacity==9.0.0
tiktoken==0.8.0
tomli==2.2.1
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
